% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean().R
\name{euclidean}
\alias{euclidean}
\title{Uses the Euclidean Algorithm to find the greatest common divisor of two integers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An integer.}

\item{b}{An integer.}
}
\value{
Returns greatest common divisor of \code{a} and \code{b}.
}
\description{
Uses the Euclidean Algorithm to find the greatest common divisor of two integers.
}
\examples{
euclidean(4, 64)
euclidean(17, 3)
euclidean(16, 16)
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}

