% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Graph data from Wikipedia}
\format{A data frame with 18 rows and 3 variables:
\describe{
   \item{v1}{first edge of the graph}
   \item{v2}{second edge of the graph}
   \item{w}{weight of the edge}
}}
\usage{
wiki_graph
}
\description{
A data set containing the edges and edge weights of the first graph on Dijkstra's algorithm's Wikipedia page
}
\references{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\keyword{datasets}

